Updated config.h
changed Resolution_X and Resolution_Y from 1920p to 800x600.
This is a band aid fix that forcefully scales the  game resolution. This fix seems that it could bring us problems when alt-tabbing or when changing full screen while the game is in full motion.

After we are done with the game fundamentals, I'll continue to look into other options. Rn we got our aims on one of LGF's PRs.