Fixed bug regarding card pickers
In order to unify the data and bitmap parts of a Seed card, I created a map containing both: that way we can more easily comment any info out and the whole seed card is out.

I also changed the hard-coded price in the SeedCards data to use the one hard coded over Plants.h for easier refactoring

I did also refactor the reading card part in Map::CreatePlantOnClick() so  that the order aligns with the order of the seed cards appearances for quicker refactoring.